Check if Array ascending
wordCount that returns a SortedMap<String, Integer> where keys are all different strings and values are the number of times that string appears in the array. The method takes an array of strings as input;
printMap that prints all entries of the map ("key : value").
wordCount that returns a SortedMap<String, Integer> where keys are all different strings and values are the number of times that string appears in the array. The method takes an array of strings as input;
printMap that prints all entries of the map ("key : value").