number occurrence using java
This program is made using java. The program is about counting occurrence of a particular digit in a given number.
For example if we input 45532 and we want to count the number of times 5 occurs then the output would be 2.

CODE EXPLANATION:
suppose the user enters 45532 (this number gets stored in 'n')
and if the user want to find number of times 5 occurs. (5 gets stored in 'a')
-->initially count = 0
Entering while loop:
rem = n % 10 (in this case: 45532 % 10 = 2 ,so rem = 2)
now rem is not equal to a(5 in this case) so count will remain 0 and will not get incremented.
-->n = n / 10 (45532 / 10 = 4553)

the while loop will run till n>0.
so at the end we will get output as : 2.